refactor to support custom implementations of tegMetric
